Reorganised controller package
