+++ src/main/java/com/codecool/shop/controller/event_controller/postal_time_service_controller/PostalTimeServiceController.java
@@ -1,4 +1,4 @@
-package com.codecool.shop.controller.postal_time_service_controller;
+package com.codecool.shop.controller.event_controller.postal_time_service_controller;
 
 import com.codecool.shop.dao.implementation.db.ShippingDataDB;
 import com.codecool.shop.model.Order;
