+++ src/main/java/com/codecool/shop/controller/rendered_page/UserController.java
@@ -1,4 +1,4 @@
-package com.codecool.shop.controller;
+package com.codecool.shop.controller.rendered_page;
 
 
 import com.codecool.shop.dao.implementation.db.UserLoginHandlerDAODB;
