--- src/main/java/com/codecool/shop/dao/implementation/mem/OrderDaoMem.java
@@ -1,55 +1,0 @@
-package com.codecool.shop.dao.implementation.mem;
-
-import com.codecool.shop.dao.OrderDao;
-import com.codecool.shop.model.Order;
-import com.codecool.shop.model.Status;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Collectors;
-
-// Creating a data storing structure for storing the orders
-public class OrderDaoMem implements OrderDao{
-
-    private static List<Order> DATA = new ArrayList<>();
-    private static OrderDaoMem instance = null;
-
-    private OrderDaoMem() {
-    }
-    // Using the Singleton pattern, prevent it from  instantiating multiple times
-    public static OrderDaoMem getInstance() {
-        if (instance == null) {
-            instance = new OrderDaoMem();
-        }
-        return instance;
-    }
-
-    @Override
-    public void add(Order order) {
-        DATA.add(order);
-    }
-
-    @Override
-    public Order find(int id) {
-        return DATA.stream().filter(t -> t.getId() == id).findFirst().orElse(null);
-    }
-
-    @Override
-    public void remove(int id) {
-        DATA.remove(find(id));
-    }
-
-    @Override
-    public List<Order> getAll() {
-        return DATA;
-    }
-
-    @Override
-    public List<Order> getBy(Status status) {
-        return DATA.stream().filter(t -> t.getStatus().equals(status)).collect(Collectors.toList());
-    }
-
-    void clear(){
-        DATA = new ArrayList<>();
-    }
-}
