--- src/main/java/com/codecool/shop/dao/implementation/mem/ProductCategoryDaoMem.java
@@ -1,51 +1,0 @@
-package com.codecool.shop.dao.implementation.mem;
-
-
-import com.codecool.shop.dao.ProductCategoryDao;
-import com.codecool.shop.model.ProductCategory;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class ProductCategoryDaoMem implements ProductCategoryDao {
-
-    private static List<ProductCategory> DATA = new ArrayList<>();
-    private static ProductCategoryDaoMem instance = null;
-
-    /* A private Constructor prevents any other class from instantiating.
-     */
-    private ProductCategoryDaoMem() {
-    }
-
-    public static ProductCategoryDaoMem getInstance() {
-        if (instance == null) {
-            instance = new ProductCategoryDaoMem();
-        }
-        return instance;
-    }
-
-    @Override
-    public void add(ProductCategory category) {
-        category.setId(DATA.size() + 1);
-        DATA.add(category);
-    }
-
-    @Override
-    public ProductCategory find(int id) {
-        return DATA.stream().filter(t -> t.getId() == id).findFirst().orElse(null);
-    }
-
-    @Override
-    public void remove(int id) {
-        DATA.remove(find(id));
-    }
-
-    @Override
-    public List<ProductCategory> getAll() {
-        return DATA;
-    }
-
-    void clear(){
-        DATA = new ArrayList<>();
-    }
-}
