--- src/main/java/com/codecool/shop/dao/implementation/mem/ProductDaoMem.java
@@ -1,64 +1,0 @@
-package com.codecool.shop.dao.implementation.mem;
-
-
-import com.codecool.shop.dao.ProductDao;
-import com.codecool.shop.model.Product;
-import com.codecool.shop.model.ProductCategory;
-import com.codecool.shop.model.Supplier;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Collectors;
-
-public class ProductDaoMem implements ProductDao {
-
-    private static List<Product> DATA = new ArrayList<>();
-    private static ProductDaoMem instance = null;
-
-    /* A private Constructor prevents any other class from instantiating.
-     */
-    private ProductDaoMem() {
-    }
-
-    public static ProductDaoMem getInstance() {
-        if (instance == null) {
-            instance = new ProductDaoMem();
-        }
-        return instance;
-    }
-
-    @Override
-    public void add(Product product) {
-        product.setId(DATA.size() + 1);
-        DATA.add(product);
-    }
-
-    @Override
-    public Product find(int id) {
-        return DATA.stream().filter(t -> t.getId() == id).findFirst().orElse(null);
-    }
-
-    @Override
-    public void remove(int id) {
-        DATA.remove(find(id));
-    }
-
-    @Override
-    public List<Product> getAll() {
-        return DATA;
-    }
-
-    @Override
-    public List<Product> getBy(Supplier supplier) {
-        return DATA.stream().filter(t -> t.getSupplier().equals(supplier)).collect(Collectors.toList());
-    }
-
-    @Override
-    public List<Product> getBy(ProductCategory productCategory) {
-        return DATA.stream().filter(t -> t.getProductCategory().equals(productCategory)).collect(Collectors.toList());
-    }
-
-    void clear(){
-        DATA = new ArrayList<>();
-    }
-}
