--- src/main/java/com/codecool/shop/dao/implementation/mem/SupplierDaoMem.java
@@ -1,50 +1,0 @@
-package com.codecool.shop.dao.implementation.mem;
-
-import com.codecool.shop.dao.SupplierDao;
-import com.codecool.shop.model.Supplier;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class SupplierDaoMem implements SupplierDao {
-
-    private static List<Supplier> DATA = new ArrayList<>();
-    private static SupplierDaoMem instance = null;
-
-    /* A private Constructor prevents any other class from instantiating.
-     */
-    private SupplierDaoMem() {
-    }
-
-    public static SupplierDaoMem getInstance() {
-        if (instance == null) {
-            instance = new SupplierDaoMem();
-        }
-        return instance;
-    }
-
-    @Override
-    public void add(Supplier category) {
-        category.setId(DATA.size() + 1);
-        DATA.add(category);
-    }
-
-    @Override
-    public Supplier find(int id) {
-        return DATA.stream().filter(t -> t.getId() == id).findFirst().orElse(null);
-    }
-
-    @Override
-    public void remove(int id) {
-        DATA.remove(find(id));
-    }
-
-    @Override
-    public List<Supplier> getAll() {
-        return DATA;
-    }
-
-    void clear(){
-        DATA = new ArrayList<>();
-    }
-}
