+++ src/main/resources/connection/connection.properties
@@ -5,7 +5,7 @@ user=postgres
 password=postgres
 
 # Email Connection Properties
-sender_email=
+sender_email= pindurpandurok.codecool@gmail.com
 host=smtp.gmail.com
-sender_password=
+sender_password= pindurpandurok
