+++ src/test/java/com/codecool/shop/dao/JunitTestSuite.java
@@ -1,10 +1,7 @@
 package com.codecool.shop.dao;
 
 
-import com.codecool.shop.dao.implementation.mem.OrderDaoMemTest;
-import com.codecool.shop.dao.implementation.mem.ProductCategoryDaoMemTest;
-import com.codecool.shop.dao.implementation.mem.ProductDaoMemTest;
-import com.codecool.shop.dao.implementation.mem.SupplierDaoMemTest;
+
 import com.codecool.shop.dao.implementation.db.DaoDBTest;
 
 import org.junit.runner.RunWith;
@@ -14,10 +11,6 @@ import org.junit.runners.Suite;
 @RunWith(Suite.class)
 
 @Suite.SuiteClasses({
-        OrderDaoMemTest.class,
-        ProductCategoryDaoMemTest.class,
-        ProductDaoMemTest.class,
-        SupplierDaoMemTest.class,
         DaoDBTest.class,
 })
